updatesklearn

+++ pgbm/sklearn/common.pyx
@@ -30,7 +30,7 @@ PREDICTOR_RECORD_DTYPE = np.dtype([
     ('value', Y_DTYPE),
     ('variance', Y_DTYPE),
     ('count', np.uint32),
-    ('feature_idx', np.uint32),
+    ('feature_idx', np.intp),
     ('num_threshold', X_DTYPE),
     ('missing_go_to_left', np.uint8),
     ('left', np.uint32),
